Various Fixes
The clear button now works as it should. Pressing it clears the
storage and then reloads the show list which appears empty. This
was fixed by emptying the table of all its elements before
reloading it.

Adding shows when none are currently in storage now works
properly. This was fixed by keeping an empty shows object in
storage.