Fixes and Updates
The issue of the show's url not being updated properly was fixed.
This was done by updating the url, saving it to storage, and then
opening the new tab rather than opening the tab first. This
functionality now works as expected.

The name of the util module was change from 'util' to 'Util'. This
was done to match with naming conventions in JavaScript.

More comments were added to better document the code.