Major Refactor
The code was completely reworked to use bootstrap.js and
handlebars.js. This was done to make the UI look much nicer and
easier to use. Bootstrap also makes the layout of the html
easier to deal with.

Templating is done using handlebars.js. It alleviates manually
creating all the DOM elements that go into each show's UI. It
is all contained within a template. The template can then be
rendered with each show's data and info. Since handlebars does not
comply with the Chrome Security Policy, it had to be sanboxed in
template.html and communicated with by sending messages through
an iframe inside popup.html.

The util.js module was removed due to it no longer being used.
Jquery was also removed because it doesn't make sense to keep 3rd
party code in version control.