Made functions generic.
The addShowFieldsEmpty, and resetAddShowFieds functions were
removed and replaced with generic fieldsEmpty and resetFields
functions. These new functions iterate through a list of field
Id's to achieve functionality. This was done to increase the
reusability of the functions.

The Show function was also updated to get its data from any three
field Id's. This was done so that the function can grab data from
any input source (either the new show form or from a shows own
info form).