Added button functionality
The next/previous buttons now work as expected. The remove buttons
remove the show from chrom storage but do not yet update the UI
to indicate show removal.

An "updateStorage" function was added to util.js. This function
updates chrome storage with the given show object replacing the
show object currently in storage. It does not add shows to
storage.

The "updateUrl" function was removed as it was not necessary.
Updating the url can be easily done inline when needed.