Started using Bower
Bower is now being used to manage the extensions dependencies.
bower.json and a .bowerrc files were added to achieve this. The
.bowerrc file makes the default directory for packages the lib
directory. This change was made so that dependencies could be
easily managed and installed. It also makes contributing to the
project easier becuase contributors won't have to download the
dependencies manually.