Removed sandboxed page, precompiled template
The handlebars show template was precompiled so that a sanboxed
page would not be necessary for the extension. This was done to
simplify the code and reduce startup time for the extension. The
template html and js files were removed. A templates directory was
added inside the js directory containing the show.handlebars
template and the compiled show.js template.