Added makefile, updated style.css
A makefile was added in order to easily build the project. Building
the project is done by copying only the essential files needed to
run the extension. All js/css files are concatenated into single
files. The library code is switched to the minified versions also.

The makefile doesn't properly build the lib.js file. Concatenating
the minified js libaries is causing an issue.

The style.css file was updated to increase the padding around the
add show buttons.