Updated makefile
The makefile was updated so that it now minifies all js files
before concatenating them. This was done to further decrease the
size of the extension when packed. The steps for building
popup.html were also updated so that only popup.js and style.css
are referenced. This was done because all js and css is concatted
into those two files. Building the directory structure was also
separated into its own dependency. This simplified the makefile.