Removed page load event listener
The event listener for the page load event was removed and the
loading shows from chrome storage now happens inside the document
ready listener. This was done because the page load listener was
only there to make the sandboxed template work properly. Since
templates are now precompiled, the lisenter is no longer necessary
and loading the shows can happen as soon as the document is ready.