Began initial backbone.js refactor.
The initial code was written in order to refactor the extension
to use bakcbone.js. The basic structure for the extension was
created. This includes a show model, a show collection, a show
view and a show list view. This was done so that the code will be
better organized, easier to maintain, and easier to reason about.