Continued backbone refactor
The show model data is now being retrived from chrome storage.
Persisting updated data has yet to be implemented.

The new show functionality was implemented and works as expected.

The previouse, next, and update functionality for each show has
been implemented and works as expected.

Unecessary ID's were removed from the show model due to backbone
being able to locally scope jquery class selectors. The only ID's
left are the infoId and removeId. These are only used for the
bootstrap collapse functionality.

The way the global app object is declared was changed. It is now
passed in to each IIFE as the 'exports' object. This makes it
obvious what is public and what is private to the closure.

An App.Util module was added in util.js. This will serve as a
module for any utility or misc functionality.