Continued Backbone refactor
The makefile was updated to support the addition/removal of files
for the refactor.

The sync method for the Show collection was overriden in order to
support retrieving/storing data to chrome storage. This
functionality is not yet complete. As of this commit, the fetch()
method seems to be working.