General Update
The javascript is now running in strict mode due to having the
'use strict'; string at the top of each module function. This was
done to better avoid common errors and mistakes.

The module exports object declarations were changed form
'exports = {}' to 'var exports = {};'. This was done to avoid
polluting the global namespace.

The for loops were removed from the fieldsEmpty() methods and
replaced with forEach instead. This improves readability.