Added clearFields method and added event handler
A private clearFields method was in the App.Util module that clears
the given array of jquery selector strings. This was added so that
the new show form can be cleared after a new show is added or if
the cancel button is pressed.

An event handler was added to the cancel button in the new show
form. The handler calls the clearFields method and clears the
form values.