Updated makefile, changed notation for show template
The makefile was updated so that the built extension functions as
expected. The error was that the template was being concatenated
after the other js/* files and the show view wasn't able to access
the template. The compiled template is now being added before the
js/* files and everything works as expected.

The notation for accesssing the template was also changed from
bracket notation to dot notation.