Updated App.Util module
The App.Util module was updated so that the fieldsEmpty() method
is now public and takes an extra context parameter. This was done
so that the show view could make use of its functionality. The
context argument allows the scoped jquery object to be passed in
from the show view and fields can be selected using the class
selectors. The fieldsEmpty() method in the show view was removed
due to it no longer being needed.