Fixed call stack error in views.js
The event handler for removing the show view and destroying the
model was causing an infinite recursive loop and exploding the
call stack. This was due to the event handler being named 'remove'
and also calling 'this.remove()'. The problem was fixed by renaming
the handler to '_remove', so that there is no longer a naming
conflict.