Added a no shows message
A message that displays when there are no shows currently in
storage was added. The message informs the user that there aren't
any shows to display and then directs them to use the add show
button to begin adding shows to their library.

This feature was achieved by simply hiding or showing some text
inside a div using jQuery's hide/show methods. The show list view
checks for the presence of shows and shows/hides the message
accordingly.