Switched to browserify, ditched bower and makefile
Browserify is now used to build the js files into a single script
thats included in the popup.html. This was done to simplify the
build process. It also allows better modularity as libaries and
scripts can be loaded using the require() function.

Switching to browserify also allows for libaries to be installed
through npm and not bower. this simplifies the project as all
dependencies are installed from a single package manager.

The makefile was also replaced with a simple build.sh shell script.
This was done as the makefile was not solving any issues and making
the build process more complicated than it needed to be.